Split routes into internally and externally reachable.
