+++ components/server/src/initialization/bottle.py
@@ -7,21 +7,8 @@ from routes.plugins import AuthPlugin, InjectionPlugin
 
 # isort: off
 # pylint: disable=unused-import
-from routes import (  # lgtm [py/unused-import]
-    auth,
-    changelog,
-    datamodel,
-    documentation,
-    logo,
-    measurement,
-    metric,
-    notification,
-    report,
-    reports_overview,
-    server,
-    source,
-    subject,
-)
+from routes.external import *  # lgtm [py/unused-import]
+from routes.internal import *  # lgtm [py/unused-import]
 
 # isort: on
