+++ components/server/src/routes/external/auth.py
