+++ components/server/src/routes/external/documentation.py
