+++ components/server/src/routes/external/logo.py
