+++ components/server/src/routes/external/metric.py
@@ -16,22 +16,6 @@ from server_utilities.functions import sanitize_html, uuid
 from server_utilities.type import MetricId, SubjectId
 
 
-@bottle.get("/internal-api/v3/metrics", authentication_required=False)
-def get_metrics(database: Database):
-    """Get all metrics."""
-    metrics: dict[str, Any] = {}
-    for report in latest_reports(database):
-        issue_tracker = report.get("issue_tracker", {})
-        has_issue_tracker = bool(issue_tracker.get("type") and issue_tracker.get("parameters", {}).get("url"))
-        for subject in report["subjects"].values():
-            for metric_uuid, metric in subject["metrics"].items():
-                metric["report_uuid"] = report["report_uuid"]
-                if has_issue_tracker and metric.get("issue_ids"):
-                    metric["issue_tracker"] = issue_tracker
-                metrics[metric_uuid] = metric
-    return metrics
-
-
 @bottle.post("/api/v3/metric/new/<subject_uuid>", permissions_required=[EDIT_REPORT_PERMISSION])
 def post_metric_new(subject_uuid: SubjectId, database: Database):
     """Add a new metric."""
