+++ components/server/src/routes/internal/__init__.py
@@ -1,0 +1,4 @@
+"""Internal routes."""
+
+from .measurement import post_measurement
+from .metric import get_metrics
