+++ components/server/tests/routes/external/test_datamodel.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 import bottle
 
 from database.datamodels import default_source_parameters, default_subject_attributes, insert_new_datamodel
-from routes import datamodel
+from routes.external import get_data_model
 from server_utilities.functions import md5_hash
 
 
@@ -20,19 +20,19 @@ class DataModelTest(unittest.TestCase):
     def test_get_data_model(self):
         """Test that the data model can be retrieved."""
         data_model = self.database.datamodels.find_one.return_value = dict(_id=123, timestamp="now")
-        self.assertEqual(data_model, datamodel.get_data_model(self.database))
+        self.assertEqual(data_model, get_data_model(self.database))
 
     def test_get_data_model_missing(self):
         """Test that the data model is None if it's not there."""
         self.database.datamodels.find_one.return_value = None
-        self.assertEqual({}, datamodel.get_data_model(self.database))
+        self.assertEqual({}, get_data_model(self.database))
 
     @patch("bottle.request")
     def test_get_data_model_unchanged(self, mocked_request):
         """Test that a 304 is returned when the data model is unchanged."""
         mocked_request.headers = {"If-None-Match": "W/" + md5_hash("now")}
         self.database.datamodels.find_one.return_value = dict(_id=123, timestamp="now")
-        self.assertRaises(bottle.HTTPError, datamodel.get_data_model, self.database)
+        self.assertRaises(bottle.HTTPError, get_data_model, self.database)
 
     def test_insert_data_model_with_id(self):
         """Test that a new data model can be inserted."""
