+++ components/server/tests/routes/external/test_logo.py
@@ -3,7 +3,7 @@
 import unittest
 from unittest.mock import patch
 
-from routes.logo import get_logo
+from routes.external import get_logo
 
 
 class LogoTest(unittest.TestCase):
