+++ components/server/tests/routes/external/test_metric.py
@@ -3,16 +3,9 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from routes.metric import (
-    delete_metric,
-    get_metrics,
-    post_metric_attribute,
-    post_metric_copy,
-    post_metric_new,
-    post_move_metric,
-)
+from routes.external import delete_metric, post_metric_attribute, post_metric_copy, post_metric_new, post_move_metric
 
-from ..fixtures import (
+from ...fixtures import (
     JOHN,
     METRIC_ID,
     METRIC_ID2,
@@ -399,60 +392,6 @@ class MetricTest(unittest.TestCase):
         )
         self.assert_delta(expected_description)
         self.assert_delta(expected_description, uuids=[REPORT_ID2, SUBJECT_ID2, METRIC_ID], report=target_report)
-
-    def test_get_metrics(self):
-        """Test that the metrics can be retrieved and deleted reports are skipped."""
-        self.database.reports_overviews.find_one.return_value = dict(_id="id", title="Reports", subtitle="")
-        self.database.reports.distinct.return_value = [REPORT_ID, "deleted_report"]
-        self.database.reports.find_one.side_effect = [self.report, dict(deleted=True)]
-        self.assertEqual(
-            {
-                METRIC_ID: dict(
-                    report_uuid=REPORT_ID,
-                    name="Metric",
-                    addition="sum",
-                    accept_debt=False,
-                    type="metric_type",
-                    tags=["security"],
-                    target="0",
-                    sources=dict(
-                        source_uuid=dict(
-                            name="Source", type="source_type", parameters=dict(url="https://url", password="password")
-                        )
-                    ),
-                )
-            },
-            get_metrics(self.database),
-        )
-
-    def test_get_metrics_with_issue(self):
-        """Test that the metrics can be retrieved and the issue tracker is included."""
-        self.database.reports_overviews.find_one.return_value = dict(_id="id", title="Reports", subtitle="")
-        self.database.reports.distinct.return_value = [REPORT_ID]
-        self.database.reports.find_one.side_effect = [self.report]
-        self.report["issue_tracker"] = dict(type="jira", parameters=dict(url="https://jira"))
-        self.report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["issue_ids"] = ["FOO-42"]
-        self.assertEqual(
-            {
-                METRIC_ID: dict(
-                    report_uuid=REPORT_ID,
-                    name="Metric",
-                    addition="sum",
-                    accept_debt=False,
-                    type="metric_type",
-                    tags=["security"],
-                    target="0",
-                    issue_ids=["FOO-42"],
-                    issue_tracker=dict(type="jira", parameters=dict(url="https://jira")),
-                    sources=dict(
-                        source_uuid=dict(
-                            name="Source", type="source_type", parameters=dict(url="https://url", password="password")
-                        )
-                    ),
-                )
-            },
-            get_metrics(self.database),
-        )
 
     def test_delete_metric(self):
         """Test that the metric can be deleted."""
