+++ components/server/tests/routes/external/test_notification.py
@@ -3,10 +3,14 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from routes.notification import (
-    post_new_notification_destination, delete_notification_destination, post_notification_destination_attributes)
+from routes.external import (
+    post_new_notification_destination,
+    delete_notification_destination,
+    post_notification_destination_attributes,
+)
 
-from ..fixtures import REPORT_ID, NOTIFICATION_DESTINATION_ID, create_report
+from ...fixtures import REPORT_ID, NOTIFICATION_DESTINATION_ID, create_report
+
 
 @patch("bottle.request")
 class PostNotificationAttributesTest(unittest.TestCase):
@@ -16,11 +20,13 @@ class PostNotificationAttributesTest(unittest.TestCase):
         """Define variables that are used in multiple tests."""
         self.database = Mock()
         self.report = dict(
-            _id="id", report_uuid=REPORT_ID, title="Report",
-            notification_destinations={NOTIFICATION_DESTINATION_ID: dict(
-                teams_webhook="",
-                name="notification_destination",
-                url="")})
+            _id="id",
+            report_uuid=REPORT_ID,
+            title="Report",
+            notification_destinations={
+                NOTIFICATION_DESTINATION_ID: dict(teams_webhook="", name="notification_destination", url="")
+            },
+        )
         self.database.reports.find.return_value = [self.report]
         self.database.datamodels.find_one.return_value = dict(_id="id")
         self.email = "john@example.org"
@@ -32,10 +38,14 @@ class PostNotificationAttributesTest(unittest.TestCase):
         post_notification_destination_attributes(REPORT_ID, NOTIFICATION_DESTINATION_ID, self.database)
         self.database.reports.insert.assert_called_once_with(self.report)
         self.assertEqual(
-            dict(uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID], email=self.email,
-                 description="John changed the 'name' of notification destination 'notification_destination' "
-                             "in report 'Report' from 'notification_destination' to 'new name'."),
-            self.report["delta"])
+            dict(
+                uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID],
+                email=self.email,
+                description="John changed the 'name' of notification destination 'notification_destination' "
+                "in report 'Report' from 'notification_destination' to 'new name'.",
+            ),
+            self.report["delta"],
+        )
 
     def test_post_notification_destination_unchanged_attribute(self, request):
         """Test changing the name of a notification destination."""
@@ -49,11 +59,15 @@ class PostNotificationAttributesTest(unittest.TestCase):
         post_notification_destination_attributes(REPORT_ID, NOTIFICATION_DESTINATION_ID, self.database)
         self.database.reports.insert.assert_called_once_with(self.report)
         self.assertEqual(
-            dict(uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID], email=self.email,
-                 description="John changed the 'name' and 'url' of notification destination 'notification_destination' "
-                             "in report 'Report' from 'notification_destination' and '' to 'new name' and "
-                             "'https://newurl'."),
-            self.report["delta"])
+            dict(
+                uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID],
+                email=self.email,
+                description="John changed the 'name' and 'url' of notification destination 'notification_destination' "
+                "in report 'Report' from 'notification_destination' and '' to 'new name' and "
+                "'https://newurl'.",
+            ),
+            self.report["delta"],
+        )
 
 
 class NotificationDestinationTest(unittest.TestCase):
@@ -73,9 +87,13 @@ class NotificationDestinationTest(unittest.TestCase):
         self.assertTrue(post_new_notification_destination(REPORT_ID, self.database)["ok"])
         notification_destinations_uuid = list(self.report["notification_destinations"].keys())[1]
         self.assertEqual(
-            dict(uuids=[REPORT_ID, notification_destinations_uuid], email=self.email,
-                 description="Jenny created a new destination for notifications in report 'Report'."),
-            self.report["delta"])
+            dict(
+                uuids=[REPORT_ID, notification_destinations_uuid],
+                email=self.email,
+                description="Jenny created a new destination for notifications in report 'Report'.",
+            ),
+            self.report["delta"],
+        )
 
     def test_add_first_new_notification_destination(self):
         """Test that a notification destination can be added."""
@@ -86,15 +104,24 @@ class NotificationDestinationTest(unittest.TestCase):
         self.assertTrue(post_new_notification_destination(REPORT_ID, self.database)["ok"])
         notification_destinations_uuid = list(report_without_destinations["notification_destinations"].keys())[0]
         self.assertEqual(
-            dict(uuids=[REPORT_ID, notification_destinations_uuid], email=self.email,
-                 description="Jenny created a new destination for notifications in report 'Report'."),
-            report_without_destinations["delta"])
+            dict(
+                uuids=[REPORT_ID, notification_destinations_uuid],
+                email=self.email,
+                description="Jenny created a new destination for notifications in report 'Report'.",
+            ),
+            report_without_destinations["delta"],
+        )
 
     def test_delete_notification_destination(self):
         """Test that a notification destination can be deleted."""
         self.assertEqual(
-            dict(ok=True), delete_notification_destination(REPORT_ID, NOTIFICATION_DESTINATION_ID, self.database))
+            dict(ok=True), delete_notification_destination(REPORT_ID, NOTIFICATION_DESTINATION_ID, self.database)
+        )
         self.assertEqual(
-            dict(uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID], email=self.email,
-                 description="Jenny deleted destination notification_destination from report 'Report'."),
-            self.report["delta"])
+            dict(
+                uuids=[REPORT_ID, NOTIFICATION_DESTINATION_ID],
+                email=self.email,
+                description="Jenny deleted destination notification_destination from report 'Report'.",
+            ),
+            self.report["delta"],
+        )
