+++ components/server/tests/routes/external/test_reports_overview.py
@@ -4,9 +4,9 @@ import unittest
 from unittest.mock import Mock, patch
 from routes.plugins.auth_plugin import EDIT_ENTITY_PERMISSION, EDIT_REPORT_PERMISSION
 
-from routes.reports_overview import get_reports_overview, post_reports_overview_attribute
+from routes.external import get_reports_overview, post_reports_overview_attribute
 
-from ..fixtures import METRIC_ID, SOURCE_ID
+from ...fixtures import METRIC_ID, SOURCE_ID
 
 
 class ReportsOverviewTest(unittest.TestCase):
