+++ components/server/tests/routes/external/test_server.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from routes.server import get_server, QUALITY_TIME_VERSION
+from routes.external import get_server, QUALITY_TIME_VERSION
 
 
 class ServerTest(unittest.TestCase):
